Initial: DD only version of Ceph SN benchmark
This is a dd-only (no Ceph) version of the benchmark we use for Ceph storage nodes. It only runs dd and does not require one to set up a single node Ceph cluster to test. 
Since we use XFS filesystems it relies on finding mounted volumes with XFS f/s, that are not mounted at /.